requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
Flask-Limiter>=2.0.0
marshmallow>=3.0.0
PyMySQL>=1.0.2
Werkzeug>=2.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
SQLAlchemy>=1.4.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0
pytest-subtests 